More utils clean-up
