+++ lib/streamlit/file_util.py
@@ -20,7 +20,7 @@ import io
 import os
 from pathlib import Path
 
-from streamlit import env_util, util
+from streamlit import env_util, errors
 from streamlit.string_util import is_binary_string
 
 # Configuration and credentials are stored inside the ~/.streamlit folder
@@ -77,7 +77,7 @@ def streamlit_read(path, binary=False):
     """
     filename = get_streamlit_file_path(path)
     if os.stat(filename).st_size == 0:
-        raise util.Error('Read zero byte file: "%s"' % filename)
+        raise errors.Error(f'Read zero byte file: "{filename}"')
 
     mode = "r"
     if binary:
@@ -109,13 +109,13 @@ def streamlit_write(path, binary=False):
         with open(path, mode) as handle:
             yield handle
     except OSError as e:
-        msg = ["Unable to write file: %s" % os.path.abspath(path)]
+        msg = [f"Unable to write file: {os.path.abspath(path)}"]
         if e.errno == errno.EINVAL and env_util.IS_DARWIN:
             msg.append(
                 "Python is limited to files below 2GB on OSX. "
                 "See https://bugs.python.org/issue24658"
             )
-        raise util.Error("\n".join(msg))
+        raise errors.Error("\n".join(msg))
 
 
 def get_static_dir() -> str:
