+++ lib/streamlit/util.py
@@ -19,15 +19,8 @@ from __future__ import annotations
 import dataclasses
 import functools
 import hashlib
-import os
-import subprocess
 import sys
-from typing import Any, Callable, Final
-
-from streamlit import env_util
-
-# URL of Streamlit's help page.
-HELP_DOC: Final = "https://docs.streamlit.io/"
+from typing import Any, Callable
 
 # Due to security issue in md5 and sha1, usedforsecurity
 # argument is added to hashlib for python versions higher than 3.8
@@ -47,60 +40,6 @@ def memoize(func: Callable[..., Any]) -> Callable[..., Any]:
         return result[0]
 
     return wrapped_func
-
-
-def open_browser(url: str) -> None:
-    """Open a web browser pointing to a given URL.
-
-    We use this function instead of Python's `webbrowser` module because this
-    way we can capture stdout/stderr to avoid polluting the terminal with the
-    browser's messages. For example, Chrome always prints things like "Created
-    new window in existing browser session", and those get on the user's way.
-
-    url : str
-        The URL. Must include the protocol.
-
-    """
-    # Treat Windows separately because:
-    # 1. /dev/null doesn't exist.
-    # 2. subprocess.Popen(['start', url]) doesn't actually pop up the
-    #    browser even though 'start url' works from the command prompt.
-    # Fun!
-    # Also, use webbrowser if we are on Linux and xdg-open is not installed.
-    #
-    # We don't use the webbrowser module on Linux and Mac because some browsers
-    # (ahem... Chrome) always print "Opening in existing browser session" to
-    # the terminal, which is spammy and annoying. So instead we start the
-    # browser ourselves and send all its output to /dev/null.
-
-    if env_util.IS_WINDOWS:
-        _open_browser_with_webbrowser(url)
-        return
-    if env_util.IS_LINUX_OR_BSD:
-        if env_util.is_executable_in_path("xdg-open"):
-            _open_browser_with_command("xdg-open", url)
-            return
-        _open_browser_with_webbrowser(url)
-        return
-    if env_util.IS_DARWIN:
-        _open_browser_with_command("open", url)
-        return
-
-    import platform
-
-    raise Error('Cannot open browser in platform "%s"' % platform.system())
-
-
-def _open_browser_with_webbrowser(url: str) -> None:
-    import webbrowser
-
-    webbrowser.open(url)
-
-
-def _open_browser_with_command(command: str, url: str) -> None:
-    cmd_line = [command, url]
-    with open(os.devnull, "w") as devnull:
-        subprocess.Popen(cmd_line, stdout=devnull, stderr=subprocess.STDOUT)
 
 
 def repr_(self: Any) -> str:
@@ -124,11 +63,6 @@ def repr_(self: Any) -> str:
 
     field_reprs = ", ".join(f"{field}={value!r}" for field, value in fields_vals)
     return f"{classname}({field_reprs})"
-
-
-# TODO: Move this into errors.py? Replace with StreamlitAPIException?
-class Error(Exception):
-    pass
 
 
 def calc_md5(s: bytes | str) -> str:
